front: started styling

+++ demo/src/main/java/com/example/demo/Service/UserService.java
@@ -1,13 +1,15 @@
 package com.example.demo.Service;
 
+import com.example.demo.Config.UserInfoDetails;
 import com.example.demo.Entity.DTO.AddUserDTO;
 import com.example.demo.Entity.User;
 import com.example.demo.Repository.EmployeeRepositoryInterface;
+import com.example.demo.Repository.LoanRepositoryInterface;
 import com.example.demo.Repository.UserRepositoryInterface;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,11 +20,14 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
     private final UserRepositoryInterface userRepository;
     private final EmployeeRepositoryInterface employeeRepository;
+    private final LoanRepositoryInterface loanRepository;
 
-    public UserService(UserRepositoryInterface userRepository, PasswordEncoder passwordEncoder, EmployeeRepositoryInterface employeeRepository) {
+    public UserService(UserRepositoryInterface userRepository, PasswordEncoder passwordEncoder,
+                       EmployeeRepositoryInterface employeeRepository, LoanRepositoryInterface loanRepository) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
         this.employeeRepository = employeeRepository;
+        this.loanRepository = loanRepository;
     }
     public List<User> getAllUsers() {
         return userRepository.findAll();
@@ -36,8 +41,15 @@ public class UserService {
     public List<User> getUserByLastName(String lastName) {
         return userRepository.findUsersByLastName(lastName);
     }
-    public List<User> getUserByUsername(String username) {
+    public List<User> getUserByUsernameContaining(String username) {
         return userRepository.findUsersByUsernameContaining(username);
+    }
+    public User getUserByUsername(Authentication authentication){
+        Object principal = authentication.getPrincipal();
+        if (principal instanceof UserInfoDetails uid) {
+            return userRepository.findUserByUsername(uid.getUsername());
+        }
+        return null;
     }
     public ResponseEntity<User> addUser(AddUserDTO userDTO) {
         if(employeeRepository.findEmployeeByUsernameOrEmail(userDTO.getUsername(), userDTO.getEmail()) == null) {
@@ -54,22 +66,30 @@ public class UserService {
         return ResponseEntity.badRequest().build();
     }
     public ResponseEntity<String> deleteUser(int id) {
-        //TODO if statement for checking active user loans
-        userRepository.deleteById(id);
-        return ResponseEntity.ok("User deleted successfully");
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            if(loanRepository.findAllByUser(user.get()).isEmpty()) {
+                userRepository.deleteById(id);
+            }
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
     public ResponseEntity<User> updateUser(int id, AddUserDTO userDTO) {
-        Optional<User> user = userRepository.findById(id);
-        if(user.isPresent()){
-            user.get().setFirstName(userDTO.getFirstName());
-            user.get().setLastName(userDTO.getLastName());
-            user.get().setUsername(userDTO.getUsername());
-            user.get().setPasswordHash(passwordEncoder.encode(userDTO.getPassword()));
-            user.get().setEmail(userDTO.getEmail());
-            user.get().setPhone(userDTO.getPhone());
-            User updatedUser = userRepository.save(user.get());
-            return ResponseEntity.ok(updatedUser);
+        if(employeeRepository.findEmployeeByUsernameOrEmail(userDTO.getUsername(), userDTO.getEmail()) == null) {
+            Optional<User> user = userRepository.findById(id);
+            if (user.isPresent()) {
+                user.get().setFirstName(userDTO.getFirstName());
+                user.get().setLastName(userDTO.getLastName());
+                user.get().setUsername(userDTO.getUsername());
+                user.get().setPasswordHash(passwordEncoder.encode(userDTO.getPassword()));
+                user.get().setEmail(userDTO.getEmail());
+                user.get().setPhone(userDTO.getPhone());
+                User updatedUser = userRepository.save(user.get());
+                return ResponseEntity.ok(updatedUser);
+            }
+            return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.notFound().build();
+        return ResponseEntity.badRequest().build();
     }
 }
